+++ src/lib/scenario/Container.svelte
@@ -1,86 +1,62 @@
 <script>
-    
-    import { slide } from 'svelte/transition'
-    import { nodes,show_rdrawer, node_info, visibleContainer } from '../../store/store.js';
-    export let name
-    export let container
-    
-    let units_rate = [
-        "Kbps","Mbps","Gbps"
-    ];
-    let units_delay = [
-        "ns","ms","s"
-    ]
-    let container_types = [
-        'csma', 'point_to_point'
-    ]
+  
+  import { slide } from 'svelte/transition'
+  import Basic from './containers/basic.svelte'
+  import Wifi from './containers/wifi.svelte'
+  import { nodes,show_rdrawer, node_info, visibleContainer } from '../../store/store.js';
+  export let name
+  export let container
+  
+  let container_types = [
+    '', 'csma', 'point_to_point', 'wifi'
+  ]
 
-    let isShown = false
-    let isNodeArrayVisible = false
+  let isShown = false
 
-    let check = false
+  function clickHandler() {
+    isShown = !isShown
+    if (isShown) 
+      visibleContainer.set(container.name)
+  }
 
-    function clickHandler() {
-        isShown = !isShown
-        if (isShown) 
-            visibleContainer.set(container.name)
-    }
+  function handleClosing() {
+    isShown = false
+    return false
+  }
+  // maybe ternary is necessary 
+  $: open = ($visibleContainer == container.name) ? 
+          (isShown) ? 
+            true
+          : false
+      : handleClosing()
 
-    function handleArray() {
-        isNodeArrayVisible = !isNodeArrayVisible
-    }
+  
 
-    function handleClosing() {
-        isShown = false
-        return false
-    }
-    // maybe ternary is necessary 
-    $: open = ($visibleContainer == container.name) ? 
-                    (isShown) ? 
-                        true
-                    : false
-            : handleClosing()
-
-    function debug() {
-        console.log(container);
-    }
-
-    function ContainerToNode(node_id){
-        $nodes.forEach(n => {
-            if(n.id==node_id){
-                if(!check)
-                    n.containers = n.containers.filter((value) => value !== container.name);
-                else
-                    n.containers = [...n.containers, container.name ]
-                show_rdrawer.update(_ => 'node_info');
-		        node_info.update(_ => n);
-            }
-        });
-    }
+  
 </script>
 
 <style>
 .parent {
-    display: flex;
-    flex-direction: column;
-    border-bottom: 2px solid navy;
-    padding: 10px;
+  display: flex;
+  flex-direction: column;
+  border-bottom: 2px solid navy;
+  padding: 10px;
 }
 
 .child {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
 }
 
 input  {
-    height: 20px;
+  height: 20px;
 }
 select  {
-    height: 26px;
+  height: 26px;
 }
 button {
-    width: 360px;
+  width: 360px;
 }
 
 ul {
@@ -92,81 +68,25 @@ ul {
 </style>
 
 <div class="topology_container parent">
-    <div class="child">
-        <button on:click={clickHandler}>{name}</button>
-    </div>
-    {#if open}
-    <div transition:slide>
-        Select type of node container: 
-        <select bind:value={container.type}>
-            {#each container_types as cont_type}
-                <option value={cont_type}>
-                    {cont_type}
-                </option>
-            {/each}
-        </select>
-
-        <div class="child">
-            <p>Data rate: </p>
-            <input type=number bind:value={container.data_rate.value} min=1 max=1024>
-            <select bind:value={container.data_rate.format}>
-                {#each units_rate as unit}
-                    <option value={unit}>
-                        {unit}
-                    </option>
-                {/each}
-            </select>
-        </div>
-        
-        <div class="child">
-            <p>Delay: </p>
-            <input type=number bind:value={container.delay.value} min=1 max=10000>
-            <select bind:value={container.delay.format}>
-                {#each units_delay as unit}
-                    <option value={unit}>
-                        {unit}
-                    </option>
-                {/each}
-            </select>
-        </div>
-        <div class="child">
-            <p>Adresa siete: </p>
-            <input bind:value={container.network_address} placeholder="Network adress">
-        </div>
-        <div class="child">
-            <p>Maska siete: </p>
-            <input bind:value={container.network_mask} placeholder="Network mask">
-        </div>
-        <div class="child">
-            <p>Názov siete: </p>
-            <input bind:value={container.network_name} placeholder="Network name">
-        </div>
-        
-        
-        <label>
-            <input type=checkbox bind:checked={container.log_pcap}> Log .pcap
-        </label>
-
-        <div>
-            <button on:click={handleArray}>Show nodes</button>
-            {#if isNodeArrayVisible}
-            <ul>
-                {#each $nodes as n (n.id) }
-                <li>
-                    <input type=checkbox bind:checked={check} on:change={() => ContainerToNode(n.id)}  bind:group={container.nodes} name="nodes" value={n.id}>
-                    Node #{n.id}
-                </li> 
-                {/each}
-            </ul>
-            {/if}
-        </div>
-
-        <div>
-            <button on:click={debug}>
-                #debug
-            </button>
-        </div>
-        
-    </div>
+  <div class="child">
+    <button on:click={clickHandler}>{name}</button>
+  </div>
+  
+  {#if open}
+    Select type of node container:
+    <select bind:value={container.type}>
+      {#each container_types as cont_type}
+        <option value={cont_type}>
+          {cont_type}
+        </option>
+      {/each}
+    </select>
+    {#if container.type =='csma' || container.type == 'point_to_point'}
+      <Basic bind:container={container}/>
+    {:else if container.type == 'wifi'}
+      <Wifi bind:container={container}/>
+    {:else}
     {/if}
+
+  {/if}
 </div>
